Bug #4255 fixed
git-svn-id: http://svn.vig/supra7/trunk@757 f05b830f-2085-0410-a479-8f75d39d0b34

+++ src/lib/Supra/FileStorage/FileStorage.php
@@ -948,10 +948,13 @@ class FileStorage
 			$sizes = $file->getImageSizeCollection();
 			foreach ($sizes as $size) {
 				$sizePath = $this->getImagePath($file, $size->getName());
-				$result = unlink($sizePath);
 				
-				if ( ! $result) {
-					throw new Exception\RuntimeException("Could not delete '$sizePath' from file storage");
+				if (file_exists($sizePath)) {
+					$result = unlink($sizePath);
+
+					if ( ! $result) {
+						throw new Exception\RuntimeException("Could not delete '$sizePath' from file storage");
+					}
 				}
 			}
 		}
